+++ src/System/Hapistrano.hs
@@ -19,12 +19,14 @@ module System.Hapistrano
   , pushReleaseWithoutVc
   , registerReleaseAsComplete
   , activateRelease
+  , linkToShared
   , rollback
   , dropOldReleases
   , playScript
   , playScriptLocally
     -- * Path helpers
   , releasePath
+  , sharedPath
   , currentSymlinkPath
   , tempSymlinkPath
   , ctokenPath )
@@ -161,6 +163,7 @@ setupDirs deployPath = do
   (exec . MkDir . releasesPath)  deployPath
   (exec . MkDir . cacheRepoPath) deployPath
   (exec . MkDir . ctokensPath)   deployPath
+  (exec . MkDir . sharedPath)    deployPath
 
 -- | Ensure that the specified repo is cloned and checked out on the given
 -- revision. Idempotent.
@@ -241,6 +244,28 @@ releasesPath
   :: Path Abs Dir      -- ^ Deploy path
   -> Path Abs Dir
 releasesPath deployPath = deployPath </> $(mkRelDir "releases")
+
+-- | Return the full path to the directory containing the shared files/directories.
+
+sharedPath
+  :: Path Abs Dir      -- ^ Deploy path
+  -> Path Abs Dir
+sharedPath deployPath = deployPath </> $(mkRelDir "shared")
+
+-- | Link something (file or directory) from the {deploy_path}/shared/ directory
+-- to a release
+
+linkToShared
+  :: TargetSystem -- ^ System to deploy
+  -> Path Abs Dir -- ^ Release path
+  -> Path Abs Dir -- ^ Deploy path
+  -> FilePath     -- ^ Thing to link in share
+  -> Hapistrano ()
+linkToShared configTargetSystem rpath configDeployPath thingToLink = do
+  destPath <- parseRelFile thingToLink
+  let dpath = rpath </> destPath
+      sharedPath' = sharedPath configDeployPath </> destPath
+  exec $ Ln configTargetSystem sharedPath' dpath
 
 -- | Construct path to a particular 'Release'.